refactor: reorganize tck tests
