+++ settings.gradle.kts
@@ -290,9 +290,6 @@ include(":tests:junit-base")
 
 // modules for system tests ------------------------------------------------------------------------
 include(":system-tests:bom-tests")
-include(":system-tests:dcp-tck-tests:presentation")
-include(":system-tests:dsp-compatibility-tests:compatibility-test-runner")
-include(":system-tests:dsp-compatibility-tests:connector-under-test")
 include(":system-tests:e2e-dataplane-tests:runtimes:data-plane")
 include(":system-tests:e2e-dataplane-tests:tests")
 include(":system-tests:e2e-transfer-test:control-plane")
@@ -302,8 +299,10 @@ include(":system-tests:e2e-transfer-test:signaling-data-plane")
 include(":system-tests:management-api:management-api-test-runner")
 include(":system-tests:management-api:management-api-test-runtime")
 include(":system-tests:protocol-2025-test")
-include(":system-tests:protocol-tck:tck-extension")
-include(":system-tests:protocol-tck:lib:tck-lib")
+include(":system-tests:tck:dcp-tck-tests")
+include(":system-tests:tck:dsp-tck-connector-under-test")
+include(":system-tests:tck:dsp-tck-tests")
+include(":system-tests:tck:tck-extension")
 include(":system-tests:telemetry:telemetry-test-runner")
 include(":system-tests:telemetry:telemetry-test-runtime")
 include(":system-tests:version-api:version-api-test-runner")
