--- system-tests/dsp-compatibility-tests/connector-under-test/src/main/java/org/eclipse/edc/tck/dsp/missing.txt
@@ -1,1 +1,0 @@
-the test harness extension should go here
