+++ system-tests/tck/dsp-tck-tests/src/test/resources/docker.tck.properties
