+++ system-tests/tck/tck-extension/src/main/java/org/eclipse/edc/tck/dsp/guard/StatefulEntityTriggerRegistry.java
